pyproject.toml
[tool.poetry]
name = "boba-tracker-server-python"
version = "0.1.0"
description = "GraphQL Server for boba-tracker project"
authors = ["Andy Huynh <andy.huynh312@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.61.1"
pydantic = {extras = ["dotenv"], version = "^1.6.1"}
uvicorn = "^0.12.1"
motor = "^2.3.0"
ariadne = "^0.12.0"
python-dotenv = "^0.14.0"
dnspython = "^2.0.0"
argon2-cffi = "^20.1.0"
pyhumps = "^1.6.1"
redis = "^3.5.3"
starlette-context = "^0.3.1"
itsdangerous = "^1.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
pre-commit = "^2.7.1"
mypy = "^0.790"
isort = "^5.6.4"
flakehell = "^0.6.1"
invoke = "^1.4.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.flakehell]
exclude = [".git", "__pycache__", "__init__.py", ".mypy_cache", ".pytest_cache"]
format = "colored"
max_line_length = 88
show_source = true

[tool.flakehell.plugins]
flake8-bandit = ["+*", "-S322"]
flake8-bugbear = ["+*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
flake8-darglint = ["+*"]
flake8-docstrings = ["+*"]
flake8-eradicate = ["+*"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-pytest-style = ["+*"]
flake8-spellcheck = ["+*"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pycodestyle = ["+*"]
pyflakes = ["+*"]
pylint = ["+*"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
